+++ packages/esm-ward-app/src/beds/occupied-bed.component.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { useTranslation } from 'react-i18next';
 import { Tag } from '@carbon/react';
 import { type WardBedProps } from '../ward-view/ward-bed.component';
-import WardPatientCard from '../ward-patient-card/ward-patient-card';
+import WardPatientCard from '../ward-patient-card/ward-patient-card.component';
 import styles from './occupied-bed.scss';
 
 const OccupiedBed: React.FC<WardBedProps> = ({ wardPatients, bed }) => {
