--- packages/esm-ward-app/src/ward-workspace/admission-request-card.scss
@@ -1,57 +1,0 @@
-@use '@carbon/styles/scss/type';
-@use '@carbon/styles/scss/spacing';
-@use '@carbon/colors';
-@import '~@openmrs/esm-styleguide/src/vars';
-
-.admissionRequestCard {
-  background-color: $ui-01;
-  border: 1px solid $ui-03;
-  color: $color-gray-70;
-  display: flex;
-  flex-direction: column;
-  > :not(:first-child) {
-    border-top: 1px colors.$gray-20 solid;
-  }
-}
-
-.wardPatientCardRow {
-  width: 100%;
-  padding: spacing.$spacing-04;
-}
-
-.wardPatientCardRow:empty {
-  display: none;
-}
-
-.wardPatientCardExtensionSlot {
-  display: none;
-
-  &:has(div:not(:empty)) {
-    display: block;
-  }
-}
-
-.wardPatientCardHeader {
-  @extend .dotSeparatedChildren;
-  display: flex;
-  flex-wrap: wrap;
-}
-
-.wardPatientName {
-  @include type.type-style('heading-compact-02');
-  color: $text-02;
-  &::before {
-    content: '' !important;
-  }
-}
-
-.dotSeparatedChildren {
-  > div:not(div:first-of-type):not(:has(div:empty)) {
-    display: flex;
-    align-items: center;
-    &::before {
-      content: '·';
-      padding: 0 spacing.$spacing-02;
-    }
-  }
-}
