+++ packages/esm-ward-app/src/ward-workspace/admission-request-card/admission-request-card.scss
