+++ packages/esm-ward-app/src/ward-workspace/admission-request-workspace/admission-requests-workspace.scss
