+++ packages/esm-ward-app/src/ward-workspace/patient-banner/patient-banner.component.tsx
@@ -1,9 +1,7 @@
 import React from 'react';
 import type { WardPatient } from '../../types';
-import {
-  useCurrentWardCardConfig,
-  WardPatientCardElement,
-} from '../../ward-patient-card/ward-patient-card-row.resources';
+import { WardPatientCardElement } from '../../ward-patient-card/ward-patient-card-element.component';
+import { useCurrentWardCardConfig } from '../../hooks/useCurrentWardCardConfig';
 import styles from './style.scss';
 
 const WardPatientWorkspaceBanner = (props: WardPatient) => {
